Implemented PauseJob. Working towards getting PauseJobGroupPausesNewJob integration test working.
Added DynamoTriggerGroup entity as need to keep track of paused trigger groups so when new triggers are added, these get paused. Decided to design this as a triggergroup entity with a state, even though I'll likely only create these records when the group is actually paused.

Removed the AWS code analysis dll's as Xamarin Studio was unhappy about them.