Fix bad merge.
Fix bad merge for trigger tests, still contained tests from when
trigger state was a string.

Fixed overridden operator to behave correctly when both sides are null.
Added test for this.

Added test to assert behaviour is correct when right hand side is an
object, but left hand side is null.